// src/store/types/Video/VideoInfo.ts
import { VideoFormat } from './VideoFormat';
// Generated by https://quicktype.io

export interface VideoInfo {
    asr:                  number |undefined;
    title:                string;
    start_time:           null;
    webpage_url_basename: string;
    upload_date:          string;
    channel_url:          string;
    uploader_id:          string;
    extractor:            string;
    acodec:               Acodec;
    duration:             number;
    is_live:              null;
    artist:               null;
    uploader_url:         string;
    release_year:         null;
    abr:                  number;
    description:          string;
    extractor_key:        string;
    uploader:             string;
    chapters:             null;
    height:               number;
    track:                null;
    tags:                 any[];
    format_id:            string;
    alt_title:            null;
    fps:                  number;
    requested_subtitles:  null;
    age_limit:            number;
    display_id:           string;
    release_date:         null;
    dislike_count:        number;
    playlist_index:       null;
    vcodec:               string;
    average_rating:       number;
    categories:           string[];
    annotations:          null;
    // subtitles:            AutomaticCaptions;
    id:                   string;
    thumbnail:            string;
    fulltitle:            string;
    view_count:           number;
    stretched_ratio:      null;
    series:               null;
    // automatic_captions:   AutomaticCaptions;
    webpage_url:          string;
    playlist:             null;
    requested_formats:    VideoFormat[];
    like_count:           number;
    ext:                  EXT;
    license:              null;
    _filename:            string;
    season_number:        null;
    resolution:           null;
    formats:              VideoFormat[];
    thumbnails:           Thumbnail[];
    album:                null;
    end_time:             null;
    channel_id:           string;
    format:               string;
    episode_number:       null;
    width:                number;
    vbr:                  null;
    creator:              null;
    durationFormatted:    string;
    filesize:             number | undefined;
}

export enum Acodec {
    Mp4A402 = 'mp4a.40.2',
    None = 'none',
    Opus = 'opus',
    Vorbis = 'vorbis',
}

// export interface AutomaticCaptions {}

export enum EXT {
    M4A = 'm4a',
    Mp4 = 'mp4',
    Webm = 'webm',
}
export interface DownloaderOptions {
    http_chunk_size: number;
}

export interface HTTPHeaders {
    'User-Agent':      string;
    'Accept-Encoding': AcceptEncoding;
    'Accept-Language': AcceptLanguage;
    'Accept-Charset':  AcceptCharset;
    Accept:            Accept;
}

export enum Accept {
    TextHTMLApplicationXHTMLXMLApplicationXMLQ09Q08 = 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
}

export enum AcceptCharset {
    ISO88591UTF8Q07Q07 = 'ISO-8859-1,utf-8;q=0.7,*;q=0.7',
}

export enum AcceptEncoding {
    GzipDeflate = 'gzip, deflate',
}

export enum AcceptLanguage {
    EnUsEnQ05 = 'en-us,en;q=0.5',
}

export enum Protocol {
    HTTPS = 'https',
}

export interface Thumbnail {
    id:  string;
    url: string;
}